+++ CHANGELOG.md
@@ -4,6 +4,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## [1.2.0] - 2026-08-25
+
+### Fixed
+- Fix special character encoding in the return of api endpoint. [#372](https://github.com/policy-design-lab/pdl-api/issues/372)
+
+### Added
+- Soybean policy storyboard APIs to return socioeconomic, soybean market balance and commodity demand data [#365](https://github.com/policy-design-lab/pdl-api/issues/365)
+- Add endpoint to return crop insurance commodities. [#368](https://github.com/policy-design-lab/pdl-api/issues/368)
+
+### Changed
+- Update crop insurance county level distribution endpoint to include commodity details. [#369](https://github.com/policy-design-lab/pdl-api/issues/369)
+
 ## [1.1.0] - 2026-04-09
 
 ### Added
@@ -274,6 +287,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Open API specification file. [#50](https://github.com/policy-design-lab/pdl-api/issues/50)
 
+[1.2.0]: https://github.com/policy-design-lab/pdl-api/compare/1.1.0...1.2.0
 [1.1.0]: https://github.com/policy-design-lab/pdl-api/compare/1.0.0...1.1.0
 [1.0.0]: https://github.com/policy-design-lab/pdl-api/compare/0.26.0...1.0.0
 [0.26.0]: https://github.com/policy-design-lab/pdl-api/compare/0.25.0...0.26.0